config: add TIMEOUT config option
Defaults to 45s instead of the 30s that Sinopia defaults to. This is
because the intended use-case for this module is in chained npm
registries. When registries are chained together in this fashion, the
initial request for a module may result in it being transferred between
multiple sites before getting to the client.