nginx: use proxy_store for tarballs
Since tarballs are never supposed to change and because we will
eventually want a method of pre-seeding a registry, switch to a
combination of try_files and proxy_store for the tarballs instead of
using a real cache.

Note that this _does_ have implications on disk usage of long-running
instances since infrequently used tarballs will never be removed.