nginx: initial support for static preseeding
Read input from STDIN on startup. If it is a tgz file, then assume it is
a node package as produced by `npm pack`. Read that tarball in and
extract the metadata from the 'package/package.json' contained in it.

Use this extracted metadata to fetch the upstream registry's metadata
for the given package and then augment the version-specific metadata for
version given in the tarball such that it looks like it was recently
published.

Use shasum to calculate the digest and then pre-seed the paths where
nginx will look for the package.json and package-X.Y.Z.tgz so that it
doesn't even bother looking upstream for them.